PascalCase naming, add types to package; add some types from network client
